+++ app/grails-app/services/de/laser/RefdataService.groovy
@@ -144,5 +144,6 @@ class RefdataService {
 
         checkResult.sort()
     }
+
 }